requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.0
fonttools==4.58.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleads==45.0.0
googleapis-common-protos==1.69.2
gspread==6.2.0
gspread-formatting==1.2.1
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
namex==0.1.0
networkx==3.4.2
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
platformdirs==4.3.6
priority==2.0.0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
Quart==0.20.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.6.0
tifffile==2025.5.26
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.14.2
zeep==4.3.1
